refactor: rename default.ts -> const.ts
