+++ packages/devextreme/js/__internal/grids/new/grid_core/toolbar/types.ts
@@ -1,6 +1,6 @@
 import type { Item as BaseToolbarItem } from '@js/ui/toolbar';
 
-import type { DEFAULT_TOOLBAR_ITEMS } from './defaults';
+import type { DEFAULT_TOOLBAR_ITEMS } from './const';
 
 export type DefaultToolbarItemName = typeof DEFAULT_TOOLBAR_ITEMS[number];
